fix(registrations): read reference suffixes as decimal on CockroachDB
CockroachDB parses a string with a leading zero as octal when casting it
to INT: CAST('012' AS INT) is 10 and CAST('008' AS INT) fails with
22P02. The seed read the zero-padded suffixes that way, so the first
allocation of a prefix could reuse a number, and the numbering query
already on develop fails for any prefix that has a '...-008' or '...-009'.
Cast to DECIMAL, which both engines read as base 10. The seeding test now
stores '009' too.

+++ packages/db/src/queries/registrations.ts
@@ -686,11 +686,12 @@ export function formatReferenceNumber(prefix: string, sequence: number): string 
  * it locks nothing. MAX runs over the number (a text MAX breaks at
  * '...-999' vs '...-1000'), and the digits-only filter keeps rows whose
  * prefix merely starts with ours ('26-TSHG-' vs '26-TSHG-CONGRES-') out of
- * the cast. CockroachDB forbids FOR UPDATE with aggregates; none is taken.
+ * the cast. The cast is to DECIMAL: CockroachDB parses a string with a
+ * leading zero as octal when casting to INT ('012' → 10, '008' → error).
  */
 async function maxReferenceSuffix(db: DbExecutor, prefix: string): Promise<number> {
   const res = await db.execute(sql`
-    SELECT MAX(CAST(seq AS INT)) AS max_seq FROM (
+    SELECT MAX(CAST(seq AS DECIMAL)) AS max_seq FROM (
       SELECT SUBSTRING("reference_number", CAST(${prefix.length + 1} AS INT)) AS seq
       FROM "registrations"
       WHERE "reference_number" LIKE ${`${escapeLike(prefix)}%`} ESCAPE '\\'
